fix type data

+++ app/dto/responseList.go
@@ -2,9 +2,9 @@ package dto
 
 // ResponseList ...
 type ResponseList struct {
-	Total    uint        `json:"total_data"`
-	Limit    uint        `json:"limit"`
-	Page     uint        `json:"page"`
-	LastPage uint        `json:"last_page"`
+	Total    uint64      `json:"total_data"`
+	Limit    uint64      `json:"limit"`
+	Page     uint64      `json:"page"`
+	LastPage uint64      `json:"last_page"`
 	Data     interface{} `json:"data"`
 }
